Change kaleidoscope folder
And add minor changes
