+++ kaleidoscope/Lexer.hpp
@@ -1,30 +1,16 @@
 #pragma once
 
-#include <functional>
-
-#include "Token.hpp"
+class std::string;
+class Reader;
 
 class Lexer {
 public:
-  Lexer(std::function<char()> r) : reader{r} {}
-
-  Token getNextToken() {
-    /*
-       callback for reader
-       it allows for flexibility in the lexer
-       it decouples how we obtain the characters from how we process them
-       TODO: investigate how to read utf-8 characters
-       could break API
-       auto mitigates API breaking
-       TODO: define error handling
-    */
-    lastCharacter = reader();
-    while (lastCharacter == ' ') {
-      lastCharacter = reader();
-    }
-  }
+  Lexer(Reader &r);
+  Token getNextToken();
 
 private:
-  std::function<char()> reader;
-  char lastCharacter;
+  Reader &reader; // NOTE: this is very bad!
+  int lastCharacter;
+  std::string identifierStr;
+  double valNum;
 };
