--- sandbox/lexer/CMakeLists.txt
@@ -1,7 +1,0 @@
-target_sources (sandbox_executable PRIVATE
-                                            Reader.h
-                                            Reader.cpp
-                                            Lexer.cpp
-                                            Lexer.h
-                                            Token.h
-)
