--- sandbox/lexer/Reader.h
@@ -1,21 +1,0 @@
-#ifndef READER_H
-#define READER_H
-
-#include <iostream>
-#include <fstream>
-#include <string>
-
-class Reader {
-public:
-    Reader(const std::string& filename);
-    ~Reader();
-    void readFile();
-    char readChar();
-    char peekNextChar();
-
-private:
-    std::string filename_;
-    std::ifstream file_;
-};
-
-#endif // READER_H
